moved external ontology to ftx
