+++ forte/data/readers/ag_news_reader.py
@@ -16,10 +16,11 @@ The reader that reads AG News data into Datapacks.
 """
 from typing import Iterator, Tuple
 
+from ftx.onto.ag_news import Description
+from ft.onto.base_ontology import Document, Title
 from forte.data.data_pack import DataPack
 from forte.data.base_reader import PackReader
-from ft.onto.ag_news import Description
-from ft.onto.base_ontology import Document, Title
+
 
 __all__ = [
     "AGNewsReader",
