+++ ftx/onto/__init__.py
@@ -1,0 +1,1 @@
+# ***automatically_generated***
