+++ tests/forte/data/ontology/test_outputs/ft/onto/example_ontology.py
@@ -1,5 +1,5 @@
 # ***automatically_generated***
-# ***source json:../../../../../../Documents/forte_develop/forte/tests/forte/data/ontology/test_specs/example_ontology.json***
+# ***source json:tests/forte/data/ontology/test_specs/example_ontology.json***
 # flake8: noqa
 # mypy: ignore-errors
 # pylint: skip-file
