+++ tests/forte/data/ontology/test_outputs/ft/onto/race_qa_ontology.py
@@ -1,5 +1,5 @@
 # ***automatically_generated***
-# ***source json:../../../../../../Documents/forte_develop/forte/tests/forte/data/ontology/test_specs/race_qa_onto.json***
+# ***source json:tests/forte/data/ontology/test_specs/race_qa_onto.json***
 # flake8: noqa
 # mypy: ignore-errors
 # pylint: skip-file
