+++ tests/forte/data/readers/ag_news_reader_test.py
@@ -19,7 +19,7 @@ import unittest
 from typing import Dict
 
 from forte.pipeline import Pipeline
-from ft.onto.ag_news import Description
+from ftx.onto.ag_news import Description
 from ft.onto.base_ontology import Document, Title
 
 from forte.data.readers import AGNewsReader
